+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - **decidim**: `decidim:check_locales` task not working properly. [\#1740]((https://https://github.com/decidim/decidim/pull/1740)
 - **decidim-core**: Disable HTML validations which sometimes caused unexpected form behavior. [\#1772](https://https://github.com/decidim/decidim/pull/1772)
 - **decidim-participatory_processes**: Scopes enabled checkbox being ignored in participatory process creation. [\#1762](https://https://github.com/decidim/decidim/pull/1762)
+- **decidim-proposals**: Fix incorrect button displayed when proposal creation not enabled. [\#1847](https://https://github.com/decidim/decidim/pull/1847)
 - **decidim-surveys**: Drag & drop icon incorrectly rendered in form. [\#1761](https://https://github.com/decidim/decidim/pull/1761)
 
 [Full Changelog](https://github.com/decidim/decidim/compare/v0.5.1...HEAD)